Moved core functionality into root `Mvp` directory
