+++ Assets/MVP/Instantiating/IPresenterFactory.cs
@@ -1,5 +1,3 @@
-using Mvp.Presenters;
-
 namespace Mvp.Instantiating
 {
     public interface IPresenterFactory
