--- Assets/MVP/Models.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7b7cda57c6ac01e4a82452e0edd3c684
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
