--- Assets/MVP/Presenters.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7dda41af919f1bc4b95c61c53351f81f
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
