--- Assets/MVP/Views.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5bfe6280811acf74d80c9eceb9e8e5a0
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
