+++ Assets/SampleProject/Scripts/Models/CubeModel.cs
@@ -1,4 +1,4 @@
-using Mvp.Models;
+using Mvp;
 using System;
 using UnityEngine;
