+++ Assets/SampleProject/Scripts/Presenters/UIPresenter.cs
@@ -2,7 +2,7 @@ using SampleProject.Messaging;
 using SampleProject.Models;
 using SampleProject.Views;
 using Mvp.Messaging;
-using Mvp.Presenters;
+using Mvp;
 
 namespace SampleProject.Presenters
 {
