+++ Assets/SampleProject/Scripts/Views/CubeView.cs
@@ -1,4 +1,4 @@
-using Mvp.Views;
+using Mvp; 
 using System;
 using UnityEngine;
